+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 9.5.3 (git master)
 * Add Planned Start and Scheduled Finish to project properties.
+* Corrected source of Must Finish By project property when reading XER files or P6 databases.
 * Updated to use POI 5.0.0.
 * When reading PMXML files, ensure that the activity calendar is set before calculating slack.